chore: remove useless ReadOnlyNetworkBuffer from API and merger it with NetworkBuffer
