+++ api/src/main/java/net/hypejet/jet/buffer/NetworkBuffer.java
@@ -13,7 +13,15 @@ import java.util.UUID;
  * @since 1.0
  * @author Codestech
  */
-public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
+public interface NetworkBuffer {
+    /**
+     * Reads a boolean from the buffer.
+     *
+     * @return the boolean
+     * @since 1.0
+     */
+    boolean readBoolean();
+
     /**
      * Writes a boolean to the buffer.
      *
@@ -21,6 +29,14 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
      * @since 1.0
      */
     void writeBoolean(boolean value);
+
+    /**
+     * Reads a byte from the buffer.
+     *
+     * @return the byte
+     * @since 1.0
+     */
+    byte readByte();
 
     /**
      * Writes a byte to the buffer.
@@ -31,12 +47,29 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeByte(byte value);
 
     /**
+     * Reads an unsigned byte from the buffer.
+     *
+     * @return the unsigned byte
+     * @since 1.0
+     */
+    @IntRange(from = 0, to = 255)
+    short readUnsignedByte();
+
+    /**
      * Writes an unsigned byte to the buffer.
      *
      * @param value the unsigned byte
      * @since 1.0
      */
     void writeUnsignedByte(@IntRange(from = 0, to = 255) short value);
+
+    /**
+     * Reads a short from the buffer.
+     *
+     * @return the short
+     * @since 1.0
+     */
+    short readShort();
 
     /**
      * Writes a short to the buffer.
@@ -47,12 +80,28 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeShort(short value);
 
     /**
+     * Reads an unsigned short from the buffer.
+     *
+     * @return the unsigned short
+     * @since 1.0
+     */
+    @IntRange(from = 0, to = 65535) int readUnsignedShort();
+
+    /**
      * Writes an unsigned short to the buffer.
      *
      * @param value the unsigned short
      * @since 1.0
      */
     void writeUnsignedShort(@IntRange(from = 0, to = 65535) int value);
+
+    /**
+     * Reads an integer from the buffer.
+     *
+     * @return the integer
+     * @since 1.0
+     */
+    int readInt();
 
     /**
      * Writes an integer to the buffer.
@@ -63,12 +112,28 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeInt(int value);
 
     /**
+     * Reads a long from the buffer.
+     *
+     * @return the long
+     * @since 1.0
+     */
+    long readLong();
+
+    /**
      * Writes a long to the buffer.
      *
      * @param value the long
      * @since 1.0
      */
     void writeLong(long value);
+
+    /**
+     * Reads a float from the buffer.
+     *
+     * @return the float
+     * @since 1.0
+     */
+    float readFloat();
 
     /**
      * Writes a float to the buffer.
@@ -79,6 +144,14 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeFloat(float value);
 
     /**
+     * Reads a double from the buffer.
+     *
+     * @return the double
+     * @since 1.0
+     */
+    double readDouble();
+
+    /**
      * Writes a double to the buffer.
      *
      * @param value the double
@@ -87,12 +160,28 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeDouble(double value);
 
     /**
+     * Reads a string from the buffer.
+     *
+     * @return the string
+     * @since 1.0
+     */
+    @NonNull
+    String readString();
+
+    /**
      * Writes a string to the buffer.
      *
      * @param value the string
      * @since 1.0
      */
     void writeString(@NonNull String value);
+
+    /**
+     * Reads a text component, which is serialized as an NBT tag.
+     *
+     * @return the text component
+     */
+    @NonNull Component readTextComponent();
 
     /**
      * Writes a text component, which is serialized as an NBT tag.
@@ -105,10 +194,27 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     /**
      * Reads a text component, which is serialized as a json string.
      *
+     * @return the text component
+     * @since 1.0
+     */
+    @NonNull Component readJsonTextComponent();
+
+    /**
+     * Reads a text component, which is serialized as a json string.
+     *
      * @param value the text component
      * @since 1.0
      */
     void writeJsonTextComponent(@NonNull Component value);
+
+
+    /**
+     * Reads a variable-length integer from the buffer.
+     *
+     * @return the variable-length integer
+     * @since 1.0
+     */
+    int readVarInt();
 
     /**
      * Writes a variable-length integer to the buffer.
@@ -119,12 +225,28 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeVarInt(int value);
 
     /**
+     * Reads a variable-length long from the buffer.
+     *
+     * @return the variable-length long
+     * @since 1.0
+     */
+    long readVarLong();
+
+    /**
      * Writes a variable-length long to the buffer.
      *
      * @param value the variable-length long
      * @since 1.0
      */
     void writeVarLong(long value);
+
+    /**
+     * Reads an unique id from the buffer.
+     *
+     * @return the unique id
+     * @since 1.0
+     */
+    @NonNull UUID readUniqueId();
 
     /**
      * Writes a unique id to the buffer.
@@ -135,6 +257,19 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
     void writeUniqueId(@NonNull UUID value);
 
     /**
+     * Reads an identifier from the buffer.
+     *
+     * @return the identifier
+     * @since 1.0
+     */
+    default @NonNull Key readIdentifier() {
+        String identifier = this.readString();
+        if (!Key.parseable(identifier)) throw identifierNotParseable(identifier);
+        // Use another method to create a key to avoid warning in IDE, which are already solved
+        return Key.key(identifier, Key.DEFAULT_SEPARATOR);
+    }
+
+    /**
      * Writes an identifier to the buffer.
      *
      * @param identifier the identifier
@@ -142,5 +277,9 @@ public interface NetworkBuffer extends ReadOnlyNetworkBuffer {
      */
     default void writeIdentifier(@NonNull Key identifier) {
         this.writeString(identifier.asString());
+    }
+
+    private static @NonNull RuntimeException identifierNotParseable(@NonNull String identifier) {
+        return new IllegalArgumentException("An identifier of \"" + identifier + "\" contains illegal characters");
     }
 }
